Some further clean ups
