+++ src/components/leftContent/RecipeList.tsx
@@ -1,13 +1,14 @@
 import { FC, useState, useEffect } from "react";
-import { Link, useHistory, useLocation, useRouteMatch } from "react-router-dom";
+import { useHistory, useLocation } from "react-router-dom";
 import styled from "styled-components";
-import { Button, TextField, Typography } from "@mui/material";
-import AddIcon from "@mui/icons-material/Add";
+import { TextField, Typography } from "@mui/material";
 import { useDebounce } from "@uidotdev/usehooks";
 
 import { RecipeModal } from "../RecipeModal";
 import { useRecipes } from "../../api";
 import { useQuery } from "../../useQuery";
+import { RecipeItem } from "./RecipeItem";
+import { RecipeListHeader } from "./RecipeListHeader";
 
 export const RecipeList: FC = () => {
   const query = useQuery();
@@ -23,34 +24,17 @@ export const RecipeList: FC = () => {
     history.replace({ pathname, search: params.toString() });
   }, [debouncedSearch, history, pathname]);
 
+  console.log(debouncedSearch);
   const { data: recipes } = useRecipes(debouncedSearch ? debouncedSearch : "");
 
   //TODO handle loading and error
 
   return (
     <Container>
-      <Header>
-        <TitleButton>
-          <Title>Recipes</Title>
-          <Button
-            startIcon={<AddIcon />}
-            variant="contained"
-            onClick={() => setRecipeModalOpen(true)}
-          >
-            Add recipe
-          </Button>
-        </TitleButton>
-        <SearchBox
-          id="outlined-basic"
-          label="Search recipes"
-          variant="outlined"
-          value={search}
-          onChange={(e) => setSearch(e.currentTarget.value)}
-        />
-      </Header>
+      <RecipeListHeader search={search} setSearch={setSearch} />
       <Recipes>
         {recipes?.map((recipe) => (
-          <Recipe recipe={recipe} key={recipe.id} />
+          <RecipeItem key={recipe.id} recipe={recipe} />
         ))}
       </Recipes>
       {recipeModalOpen && (
@@ -90,29 +74,6 @@ const Recipes = styled.div`
 
   :not(:last-child) {
     border-bottom: 1px solid grey;
-  }
-`;
-
-const Recipe: FC<{ recipe: any }> = ({ recipe }) => {
-  const { url } = useRouteMatch();
-  const { pathname } = useLocation();
-
-  const recipeId = pathname.split("/")[2];
-  return (
-    <Link to={`${url}/${recipe.id}`}>
-      <RecipeContainer isSelected={!!recipeId && +recipeId === recipe.id}>
-        {recipe.title}
-      </RecipeContainer>
-    </Link>
-  );
-};
-
-const RecipeContainer = styled.div<{ isSelected?: boolean }>`
-  padding: ${({ theme }) => theme.spacing(2, 1)};
-  background: ${({ theme, isSelected }) =>
-    isSelected ? theme.palette.grey["400"] : "inherit"};
-  &:hover {
-    background: ${({ theme }) => theme.palette.grey["300"]};
   }
 `;
